Improving MeanEvolutionResult comparers
