+++ src/GeneticSharp.Domain.UnitTests/Metaheuristics/MetaHeuristicTestBase.cs
@@ -11,6 +11,7 @@ using GeneticSharp.Domain.Metaheuristics.Primitives;
 using GeneticSharp.Domain.Mutations;
 using GeneticSharp.Domain.Populations;
 using GeneticSharp.Domain.Reinsertions;
+using GeneticSharp.Domain.Results;
 using GeneticSharp.Domain.Selections;
 using GeneticSharp.Domain.Terminations;
 using GeneticSharp.Extensions.Mathematic;
